pyproject.toml
[tool.poetry]
name = "dataplatform-kubernetes"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "dataplatform_kubernetes", from = "scripts" },
]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
crcmod = "^1.7"
PyYAML = "^6.0"
PyGithub = "^1.55"
Jinja2 = "^3.1.3"
google-cloud-bigquery = "^3.10.0"
google-cloud-bigquery-storage = "^2.19.1"
pyarrow = "^14.0.1"
slack-sdk = "^3.25.0"
google-cloud-secret-manager = "^2.16.4"
tenacity = "^8.2.3"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
pytest-mock = "*"
types-PyYAML = "*"
black = "*"
isort = "*"
flake8 = "*"
mypy = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
norecursedirs = [".poetry", "javascript"]

[tool.black]
line-length = 100
target-version = ['py38']
exclude = '(\.poetry|\.tox)'

[tool.isort]
profile = "black"
multi_line_output = 3
skip_glob = [".poetry/*", "javascript/*"]

[tool.mypy]
python_version = 3.8
follow_imports = "silent"
disallow_any_generics = true
strict_optional = true
check_untyped_defs = true
allow_redefinition = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
exclude = ["scripts/tests.*", ".poetry.*", "javascript"]

[[tool.mypy.overrides]]
module = [".poetry.*"]
ignore_errors = true

[tool.poetry.scripts]
dp = "dataplatform_kubernetes.cli:cli"
mysql-to-gcs = "argo_workflows.mysql_to_gcs.cli:cli"

